@@ -1,6 +1,6 @@
 {
   "name": "@sensenet/sn-app",
-  "version": "0.4.0",
+  "version": "0.5.0",
   "main": "dist/index.js",
   "files": [
     "dist",
@@ -27,18 +27,18 @@
   "homepage": "https://sensenet.com",
   "devDependencies": {
     "@types/autosuggest-highlight": "^3.1.0",
-    "@types/react": "^16.8.19",
+    "@types/react": "^16.8.23",
     "@types/react-autosuggest": "^9.3.8",
-    "@types/react-dom": "^16.8.4",
-    "@types/react-redux": "^7.0.9",
+    "@types/react-dom": "^16.8.5",
+    "@types/react-redux": "^7.1.1",
     "@types/react-responsive": "^3.0.3",
-    "@types/react-router": "^5.0.1",
+    "@types/react-router": "^5.0.3",
     "@types/react-router-dom": "^4.3.3",
-    "@types/uuid": "^3.4.4",
-    "autoprefixer": "^9.5.1",
+    "@types/uuid": "^3.4.5",
+    "autoprefixer": "^9.6.1",
     "awesome-typescript-loader": "^5.2.1",
-    "css-loader": "^3.0.0",
-    "file-loader": "^4.0.0",
+    "css-loader": "^3.1.0",
+    "file-loader": "^4.1.0",
     "git-revision-webpack-plugin": "^3.0.3",
     "html-webpack-plugin": "^3.2.0",
     "monaco-editor-webpack-plugin": "^1.7.0",
@@ -48,38 +48,38 @@
     "react-router-dom": "^5.0.0",
     "source-map-loader": "^0.2.4",
     "style-loader": "^0.23.1",
-    "ts-config-webpack-plugin": "^1.3.1",
-    "url-loader": "^2.0.1",
-    "webpack": "^4.32.2",
-    "webpack-bundle-analyzer": "^3.3.2",
-    "webpack-cli": "^3.3.2",
+    "ts-config-webpack-plugin": "^1.4.0",
+    "url-loader": "^2.1.0",
+    "webpack": "^4.38.0",
+    "webpack-bundle-analyzer": "^3.4.1",
+    "webpack-cli": "^3.3.6",
     "webpack-dev-server": "^3.4.1"
   },
   "dependencies": {
-    "@furystack/inject": "^4.0.1",
-    "@furystack/logging": "^2.0.1",
-    "@material-ui/core": "^4.0.1",
+    "@furystack/inject": "^4.0.2",
+    "@furystack/logging": "^2.0.2",
+    "@material-ui/core": "^4.3.1",
     "@material-ui/icons": "^4.0.1",
-    "@sensenet/client-core": "^2.1.0",
-    "@sensenet/client-utils": "^1.6.2",
-    "@sensenet/controls-react": "^2.8.0",
-    "@sensenet/default-content-types": "^1.2.2",
-    "@sensenet/document-viewer-react": "^1.1.2",
-    "@sensenet/list-controls-react": "^1.3.8",
-    "@sensenet/pickers-react": "^1.2.1",
+    "@sensenet/client-core": "^2.1.1",
+    "@sensenet/client-utils": "^1.6.3",
+    "@sensenet/controls-react": "^3.0.0",
+    "@sensenet/default-content-types": "^1.3.0",
+    "@sensenet/document-viewer-react": "^1.1.3",
+    "@sensenet/list-controls-react": "^1.3.9",
+    "@sensenet/pickers-react": "^1.2.2",
     "@sensenet/query": "^1.1.7",
-    "@sensenet/redux": "^5.1.10",
-    "@sensenet/repository-events": "^1.4.2",
+    "@sensenet/redux": "^5.1.11",
+    "@sensenet/repository-events": "^1.4.3",
     "autosuggest-highlight": "^3.1.1",
     "monaco-editor": "^0.17.0",
     "react": "^16.8.2",
     "react-autosuggest": "^9.4.3",
     "react-dom": "^16.8.2",
     "react-markdown": "^4.0.8",
-    "react-monaco-editor": "^0.26.1",
+    "react-monaco-editor": "^0.28.0",
     "react-redux": "^7.0.3",
     "react-responsive": "^7.0.0",
-    "redux": "^4.0.1",
+    "redux": "^4.0.4",
     "redux-di-middleware": "^4.0.1",
     "semaphore-async-await": "^1.5.1",
     "uuid": "^3.3.2"
