@@ -69,47 +69,47 @@
     "react-responsive": "^7.0.0",
     "react-router": "^5.0.0",
     "react-router-dom": "^5.0.0",
-    "redux": "^4.0.1",
+    "redux": "^4.0.4",
     "redux-di-middleware": "^4.0.1",
     "reflect-metadata": "^0.1.12",
     "semaphore-async-await": "^1.5.1",
     "ts-keycode-enum": "^1.0.6",
-    "typeface-roboto": "^0.0.54",
+    "typeface-roboto": "^0.0.75",
     "uuid": "^3.3.2"
   },
   "devDependencies": {
     "@cypress/webpack-preprocessor": "^4.0.3",
-    "@types/chance": "^1.0.4",
-    "@types/jest": "^24.0.9",
+    "@types/chance": "^1.0.5",
+    "@types/jest": "^24.0.16",
     "@types/lodash.groupby": "^4.6.6",
-    "@types/node": "^12.0.3",
-    "@types/react": "^16.8.19",
+    "@types/node": "^12.6.8",
+    "@types/react": "^16.8.23",
     "@types/react-custom-scrollbars": "^4.0.5",
-    "@types/react-dom": "^16.8.4",
-    "@types/react-google-recaptcha": "^1.0.0",
+    "@types/react-dom": "^16.8.5",
+    "@types/react-google-recaptcha": "^1.1.0",
     "@types/react-loadable": "^5.4.2",
-    "@types/react-redux": "^7.0.9",
+    "@types/react-redux": "^7.1.1",
     "@types/react-responsive": "^3.0.2",
     "@types/react-router-dom": "^4.3.3",
     "@types/redux-mock-store": "^1.0.1",
-    "@types/uuid": "^3.4.4",
+    "@types/uuid": "^3.4.5",
     "awesome-typescript-loader": "^5.2.1",
     "chance": "^1.0.18",
-    "copy-webpack-plugin": "^5.0.3",
+    "copy-webpack-plugin": "^5.0.4",
     "cross-env": "^5.2.0",
-    "cypress": "^3.3.1",
+    "cypress": "^3.4.1",
     "path-to-regexp": "3.0.0",
-    "raw-loader": "^3.0.0",
+    "raw-loader": "^3.1.0",
     "redux-mock-store": "^1.5.3",
     "source-map-loader": "^0.2.4",
     "style-loader": "^0.23.1",
-    "svg-url-loader": "^2.3.2",
+    "svg-url-loader": "^3.0.0",
     "uglify-es": "^3.3.9",
     "uglifyjs-webpack-plugin": "^2.1.3",
-    "url-loader": "^2.0.1",
-    "webpack": "^4.32.2",
-    "webpack-bundle-analyzer": "^3.3.2",
-    "webpack-cli": "^3.3.2",
+    "url-loader": "^2.1.0",
+    "webpack": "^4.38.0",
+    "webpack-bundle-analyzer": "^3.4.1",
+    "webpack-cli": "^3.3.6",
     "webpack-dev-server": "^3.4.1"
   }
 }
