@@ -1,6 +1,6 @@
 {
   "name": "sn-react-redux-todo-app",
-  "version": "2.0.8",
+  "version": "2.0.9",
   "description": "Todo app with React and Redux built upon Sense/Net ECM",
   "main": "index.tsx",
   "private": true,
@@ -30,31 +30,31 @@
     }
   },
   "dependencies": {
-    "@material-ui/core": "^4.0.1",
+    "@material-ui/core": "^4.3.1",
     "@material-ui/icons": "^4.0.1",
-    "@sensenet/authentication-google": "^2.0.9",
-    "@sensenet/client-core": "^2.1.0",
-    "@sensenet/controls-react": "^2.8.0",
-    "@sensenet/redux": "^5.1.10",
+    "@sensenet/authentication-google": "^2.0.10",
+    "@sensenet/client-core": "^2.1.1",
+    "@sensenet/controls-react": "^3.0.0",
+    "@sensenet/redux": "^5.1.11",
     "material-ui-pickers": "^2.0.4",
     "react": "^16.8.4",
     "react-dom": "^16.8.4",
     "react-quill": "^1.3.3",
     "react-router-redux": "^4.0.8",
-    "redux": "^4.0.1",
+    "redux": "^4.0.4",
     "redux-di-middleware": "^4.0.1"
   },
   "devDependencies": {
-    "@types/react": "^16.8.19",
-    "@types/react-dom": "^16.8.4",
-    "@types/react-redux": "^7.0.9",
+    "@types/react": "^16.8.23",
+    "@types/react-dom": "^16.8.5",
+    "@types/react-redux": "^7.1.1",
     "@types/react-router-dom": "^4.3.3",
     "@types/redux-logger": "^3.0.6",
     "awesome-typescript-loader": "^5.2.1",
-    "copy-webpack-plugin": "^5.0.3",
+    "copy-webpack-plugin": "^5.0.4",
     "source-map-loader": "^0.2.4",
-    "webpack": "^4.32.2",
-    "webpack-cli": "^3.3.2",
+    "webpack": "^4.38.0",
+    "webpack-cli": "^3.3.6",
     "webpack-dev-server": "^3.4.1"
   }
 }
