@@ -1,6 +1,6 @@
 {
   "name": "@sensenet/search-react",
-  "version": "1.2.6",
+  "version": "1.2.7",
   "description": "Search component in react with sensenet",
   "main": "dist/index.js",
   "typings": "./dist/index.d.ts",
@@ -27,10 +27,10 @@
   "license": "GPL-2.0",
   "homepage": "https://sensenet.com",
   "dependencies": {
-    "@material-ui/core": "^4.0.1",
-    "@sensenet/client-core": "^2.1.0",
-    "@sensenet/control-mapper": "^1.1.4",
-    "@sensenet/default-content-types": "^1.2.2",
+    "@material-ui/core": "^4.3.1",
+    "@sensenet/client-core": "^2.1.1",
+    "@sensenet/control-mapper": "^2.0.0",
+    "@sensenet/default-content-types": "^1.3.0",
     "@sensenet/icons-react": "^1.2.11",
     "@sensenet/query": "^1.1.7",
     "autosuggest-highlight": "^3.1.1",
@@ -40,23 +40,23 @@
   },
   "devDependencies": {
     "@types/autosuggest-highlight": "3.1.0",
-    "@types/react": "^16.8.19",
+    "@types/react": "^16.8.23",
     "@types/react-autosuggest": "^9.3.8",
-    "@types/react-dom": "^16.8.4",
+    "@types/react-dom": "^16.8.5",
     "@types/react-router-dom": "^4.3.3",
-    "autoprefixer": "^9.5.1",
+    "autoprefixer": "^9.6.1",
     "awesome-typescript-loader": "^5.2.1",
-    "copy-webpack-plugin": "^5.0.3",
-    "css-loader": "^3.0.0",
-    "file-loader": "^4.0.0",
+    "copy-webpack-plugin": "^5.0.4",
+    "css-loader": "^3.1.0",
+    "file-loader": "^4.1.0",
     "postcss-flexbugs-fixes": "^4.1.0",
     "postcss-loader": "^3.0.0",
     "source-map-loader": "^0.2.4",
     "style-loader": "^0.23.1",
-    "url-loader": "^2.0.1",
-    "webpack": "^4.32.2",
-    "webpack-bundle-analyzer": "^3.3.2",
-    "webpack-cli": "^3.3.2",
+    "url-loader": "^2.1.0",
+    "webpack": "^4.38.0",
+    "webpack-bundle-analyzer": "^3.4.1",
+    "webpack-cli": "^3.3.6",
     "webpack-dev-server": "^3.4.1"
   }
 }
