@@ -1,16 +1,14 @@
 /**
  * @module ViewControls
  */
-import React, { Component, ComponentType, createElement, ReactElement } from 'react'
+import React, { createElement, ReactElement, useState } from 'react'
 import Button from '@material-ui/core/Button'
 import Grid from '@material-ui/core/Grid'
 import Typography from '@material-ui/core/Typography'
 import { Repository } from '@sensenet/client-core'
-import { ControlSchema } from '@sensenet/control-mapper'
 import { ContentType } from '@sensenet/default-content-types'
 import MediaQuery from 'react-responsive'
 import { reactControlMapper } from '../ReactControlMapper'
-import { ReactClientFieldSetting } from '../fieldcontrols/ClientFieldSetting'
 
 /**
  * Interface for NewView properties
@@ -23,17 +21,9 @@ export interface NewViewProps {
   contentTypeName: string
   handleCancel?: () => void
   submitCallback?: () => void
-  title?: string
+  showTitle?: boolean
   extension?: string
   uploadFolderpath?: string
-}
-/**
- * Interface for NewView state
- */
-export interface NewViewState {
-  schema: ControlSchema<ComponentType, ComponentType<ReactClientFieldSetting>>
-  content: ContentType
-  controlMapper: ReturnType<typeof reactControlMapper>
 }
 
 /**
@@ -44,97 +34,68 @@ export interface NewViewState {
  *  <NewView content={content} onSubmit={createSubmitClick} />
  * ```
  */
-export class NewView extends Component<NewViewProps, NewViewState> {
-  constructor(props: any) {
-    super(props)
-    const controlMapper = reactControlMapper(this.props.repository)
-    this.state = {
-      schema: controlMapper.getFullSchemaForContentType(this.props.contentTypeName, 'new'),
-      // eslint-disable-next-line @typescript-eslint/no-object-literal-type-assertion
-      content: {} as ContentType,
-      controlMapper,
-    }
+export const NewView: React.FC<NewViewProps> = props => {
+  const controlMapper = reactControlMapper(props.repository)
+  const schema = controlMapper.getFullSchemaForContentType(props.contentTypeName, 'new')
+  const [content, setContent] = useState({})
 
-    this.handleInputChange = this.handleInputChange.bind(this)
+  const handleSubmit = (event: React.FormEvent<HTMLFormElement>) => {
+    event.preventDefault()
+    props.onSubmit && props.onSubmit(props.path, content as any, schema.schema.ContentTypeName)
+    props.submitCallback && props.submitCallback()
   }
 
-  public handleInputChange(field: string, value: unknown) {
-    this.setState({
-      content: { ...this.state.content, [field]: value },
-    })
+  const handleInputChange = (field: string, value: unknown) => {
+    setContent({ ...content, [field]: value })
   }
 
-  public render() {
-    const { onSubmit, path, title, submitCallback } = this.props
-    const { schema } = this.state
-    return (
-      <form
-        style={{ margin: '0 auto' }}
-        onSubmit={e => {
-          e.preventDefault()
-          if (onSubmit) {
-            onSubmit(path, this.state.content, schema.schema.ContentTypeName)
-          }
-          if (submitCallback) {
-            submitCallback()
-          }
-        }}>
-        {title !== undefined ? (
-          <Typography variant="h5" gutterBottom={true}>
-            {`New ${schema.schema.DisplayName}`}
-          </Typography>
-        ) : (
-          title
-        )}
-        <Grid container={true} spacing={2}>
-          {this.state.schema.fieldMappings.map(field => {
-            return (
-              <Grid
-                item={true}
-                xs={12}
-                sm={12}
-                md={field.fieldSettings.Name === 'LongTextFieldSetting' ? 12 : 6}
-                lg={field.fieldSettings.Name === 'LongTextFieldSetting' ? 12 : 6}
-                xl={field.fieldSettings.Name === 'LongTextFieldSetting' ? 12 : 6}
-                key={field.fieldSettings.Name}>
-                {createElement(
-                  this.state.controlMapper.getControlForContentField(
-                    this.props.contentTypeName,
-                    field.fieldSettings.Name,
-                    'new',
-                  ),
-                  {
-                    actionName: 'new',
-                    settings: field.fieldSettings,
-                    repository: this.props.repository,
-                    renderIcon: this.props.renderIcon,
-                    fieldOnChange: this.handleInputChange,
-                    extension: this.props.extension,
-                    uploadFolderPath: this.props.uploadFolderpath,
-                  },
-                )}
-              </Grid>
-            )
-          })}
-          <Grid item={true} xs={12} sm={12} md={12} lg={12} xl={12} style={{ textAlign: 'right' }}>
-            <MediaQuery minDeviceWidth={700}>
-              {matches =>
-                matches ? (
-                  <Button
-                    color="default"
-                    style={{ marginRight: 20 }}
-                    onClick={() => this.props.handleCancel && this.props.handleCancel()}>
-                    Cancel
-                  </Button>
-                ) : null
-              }
-            </MediaQuery>
-            <Button type="submit" variant="contained" color="secondary">
-              Submit
+  return (
+    <form style={{ margin: '0 auto' }} onSubmit={handleSubmit}>
+      {props.showTitle && (
+        <Typography variant="h5" gutterBottom={true}>
+          {`New ${schema.schema.DisplayName}`}
+        </Typography>
+      )}
+      <Grid container={true} spacing={2}>
+        {schema.fieldMappings.map(field => {
+          return (
+            <Grid
+              item={true}
+              xs={12}
+              sm={12}
+              md={field.fieldSettings.Name === 'LongTextFieldSetting' ? 12 : 6}
+              lg={field.fieldSettings.Name === 'LongTextFieldSetting' ? 12 : 6}
+              xl={field.fieldSettings.Name === 'LongTextFieldSetting' ? 12 : 6}
+              key={field.fieldSettings.Name}>
+              {createElement(
+                controlMapper.getControlForContentField(props.contentTypeName, field.fieldSettings.Name, 'new'),
+                {
+                  actionName: 'new',
+                  settings: field.fieldSettings,
+                  repository: props.repository,
+                  renderIcon: props.renderIcon,
+                  fieldOnChange: handleInputChange,
+                  extension: props.extension,
+                  uploadFolderPath: props.uploadFolderpath,
+                },
+              )}
+            </Grid>
+          )
+        })}
+        <Grid item={true} xs={12} sm={12} md={12} lg={12} xl={12} style={{ textAlign: 'right' }}>
+          <MediaQuery minDeviceWidth={700}>
+            <Button
+              color="default"
+              style={{ marginRight: 20 }}
+              onClick={() => props.handleCancel && props.handleCancel()}>
+              Cancel
             </Button>
-          </Grid>
+          </MediaQuery>
+          <Button type="submit" variant="contained" color="secondary">
+            Submit
+          </Button>
         </Grid>
-      </form>
-    )
-  }
+      </Grid>
+    </form>
+  )
 }
