@@ -920,10 +920,11 @@ export const customSchema = [
         Compulsory: false,
         OutputMethod: 0,
         Visible: false,
-        VisibleBrowse: 1,
-        VisibleEdit: 1,
-        VisibleNew: 1,
+        VisibleBrowse: 0,
+        VisibleEdit: 0,
+        VisibleNew: 0,
         DefaultOrder: 0,
+        Step: 0.5,
         Type: 'NumberFieldSetting',
       },
       {
