@@ -6,10 +6,11 @@ import FormHelperText from '@material-ui/core/FormHelperText'
 import InputLabel from '@material-ui/core/InputLabel'
 import List from '@material-ui/core/List'
 import Typography from '@material-ui/core/Typography'
-import { GenericContent, ReferenceFieldSetting, User } from '@sensenet/default-content-types'
+import { ReferenceFieldSetting, User } from '@sensenet/default-content-types'
 import React, { Component } from 'react'
 import { renderIconDefault } from '../icon'
 import { ReactClientFieldSetting } from '../ClientFieldSetting'
+import { changeJScriptValue } from '../../helpers'
 import { AvatarPicker } from './AvatarPicker'
 import { DefaultAvatarTemplate } from './DefaultAvatarTemplate'
 
@@ -46,17 +47,15 @@ const DEFAULT_AVATAR_PATH = '/Root/Sites/Default_Site/demoavatars/Admin.png'
 export interface AvatarState {
   fieldValue: any
   pickerIsOpen: boolean
-  selected?: GenericContent
+  selected?: User
 }
 
 export class Avatar extends Component<ReactClientFieldSetting<ReferenceFieldSetting, User>, AvatarState> {
   state: AvatarState = {
     fieldValue:
-      this.props.content &&
-      this.props.content[this.props.settings.Name] &&
-      this.props.content[this.props.settings.Name].Url
-        ? this.props.content[this.props.settings.Name].Url
-        : this.props.settings.DefaultValue || '',
+      (this.props.fieldValue && (this.props.fieldValue as any).Url) ||
+      changeJScriptValue(this.props.settings.DefaultValue) ||
+      '',
     pickerIsOpen: false,
     selected: undefined,
   }
@@ -97,7 +96,7 @@ export class Avatar extends Component<ReactClientFieldSetting<ReferenceFieldSett
     this.handleDialogClose()
   }
 
-  public selectItem = (content: GenericContent) => {
+  public selectItem = (content: User) => {
     this.setState({
       selected: content,
     })
@@ -129,7 +128,7 @@ export class Avatar extends Component<ReactClientFieldSetting<ReferenceFieldSett
               style={this.state.fieldValue.length > 0 ? styles.listContainer : { ...styles.listContainer, width: 200 }}>
               {
                 <DefaultAvatarTemplate
-                  repositoryUrl={this.props.repository!.configuration.repositoryUrl}
+                  repositoryUrl={this.props.repository && this.props.repository.configuration.repositoryUrl}
                   add={this.addItem}
                   actionName={this.props.actionName}
                   readOnly={this.props.settings.ReadOnly}
@@ -152,10 +151,11 @@ export class Avatar extends Component<ReactClientFieldSetting<ReferenceFieldSett
                   path={
                     this.props.uploadFolderPath ||
                     (this.props.settings.SelectionRoots && this.props.settings.SelectionRoots[0]) ||
+                    this.state.fieldValue.substring(0, this.state.fieldValue.lastIndexOf('/')) ||
                     ''
                   }
                   repository={this.props.repository!}
-                  select={content => this.selectItem(content)}
+                  select={this.selectItem}
                   renderIcon={this.props.renderIcon ? this.props.renderIcon : renderIconDefault}
                 />
                 <DialogActions>
@@ -170,7 +170,7 @@ export class Avatar extends Component<ReactClientFieldSetting<ReferenceFieldSett
         )
       case 'browse':
       default:
-        return this.props.content && this.props.content[this.props.settings.Name] ? (
+        return this.props.fieldValue ? (
           <FormControl style={styles.root as any}>
             <InputLabel shrink={true} htmlFor={this.props.settings.Name}>
               {this.props.settings.DisplayName}
@@ -178,14 +178,11 @@ export class Avatar extends Component<ReactClientFieldSetting<ReferenceFieldSett
             <List
               dense={true}
               style={
-                this.props.content[this.props.settings.Name].Url
-                  ? styles.listContainer
-                  : { ...styles.listContainer, width: 200 }
+                (this.props.fieldValue as any).Url ? styles.listContainer : { ...styles.listContainer, width: 200 }
               }>
               <DefaultAvatarTemplate
-                repositoryUrl={this.props.repository!.configuration.repositoryUrl}
-                url={this.props.content[this.props.settings.Name].Url}
-                add={this.addItem}
+                repositoryUrl={this.props.repository && this.props.repository.configuration.repositoryUrl}
+                url={(this.props.fieldValue as any).Url}
                 actionName="browse"
                 renderIcon={this.props.renderIcon ? this.props.renderIcon : renderIconDefault}
               />
