@@ -3,13 +3,13 @@
  */
 import ClickAwayListener from '@material-ui/core/ClickAwayListener'
 import FormControl from '@material-ui/core/FormControl'
-import FormHelperText from '@material-ui/core/FormHelperText'
 import Icon from '@material-ui/core/Icon'
 import InputAdornment from '@material-ui/core/InputAdornment'
 import TextField from '@material-ui/core/TextField'
-import React, { Component } from 'react'
-import { SketchPicker } from 'react-color'
+import React, { useState } from 'react'
+import { ColorResult, SketchPicker } from 'react-color'
 import { ColorFieldSetting } from '@sensenet/default-content-types'
+import { changeJScriptValue } from '../helpers'
 import { ReactClientFieldSetting } from './ClientFieldSetting'
 
 const style = {
@@ -28,140 +28,76 @@ const renderIconDefault = (name: string, color: string) => {
 }
 
 /**
- * Interface for ColorPicker state
- */
-export interface ColorPickerState {
-  value: string
-  pickerIsOpen: boolean
-}
-/**
  * Field control that represents a Color field. Available values will be populated from the FieldSettings.
  */
-export class ColorPicker extends Component<ReactClientFieldSetting<ColorFieldSetting>, ColorPickerState> {
-  /**
-   * constructor
-   * @param {object} props
-   */
-  constructor(props: ColorPicker['props']) {
-    super(props)
-    /**
-     * @type {object}
-     * @property {string} value input value
-     */
-    this.state = {
-      value: (this.props.content && this.setValue(this.props.content[this.props.settings.Name])) || '',
-      pickerIsOpen: false,
-    }
+export const ColorPicker: React.FC<ReactClientFieldSetting<ColorFieldSetting>> = props => {
+  const [value, setValue] = useState(props.fieldValue || changeJScriptValue(props.settings.DefaultValue) || '')
+  const [isPickerOpen, setIsPickerOpen] = useState(false)
 
-    this.handleChange = this.handleChange.bind(this)
-    this.openPicker = this.openPicker.bind(this)
-    this.closePicker = this.closePicker.bind(this)
+  const handleChange = (color: ColorResult) => {
+    props.fieldOnChange && props.fieldOnChange(props.settings.Name, color.hex)
+    setValue(color.hex)
   }
-  /**
-   * convert incoming default value string to proper format
-   * @param {string} value
-   */
-  public setValue(value: string) {
-    if (value) {
-      return value.replace(/<[^>]*>/g, '')
-    } else {
-      if (this.props.settings.DefaultValue) {
-        return this.props.settings.DefaultValue
-      } else {
-        return '#016d9e'
-      }
-    }
-  }
-  /**
-   * handle change event on an input
-   * @param {SytheticEvent} event
-   */
-  public handleChange(color: any) {
-    this.props.fieldOnChange && this.props.fieldOnChange(this.props.settings.Name, color.hex)
-    this.setState({ value: color.hex })
-  }
-  public openPicker() {
-    this.setState({
-      pickerIsOpen: true,
-    })
-  }
-  public closePicker() {
-    this.setState({
-      pickerIsOpen: false,
-    })
-  }
-  /**
-   * render
-   * @return {ReactElement} markup
-   */
-  public render() {
-    switch (this.props.actionName) {
-      case 'edit':
-      case 'new':
-        return (
-          <FormControl>
-            <TextField
-              label={this.props.settings.DisplayName}
-              type="text"
-              name={this.props.settings.Name}
-              id={this.props.settings.Name}
-              required={this.props.settings.Compulsory}
-              disabled={this.props.settings.ReadOnly}
-              value={this.state.value}
-              onClick={this.openPicker}
-              InputProps={{
-                startAdornment: (
-                  <InputAdornment position="start">
-                    {this.props.renderIcon
-                      ? this.props.renderIcon('lens')
-                      : renderIconDefault('lens', this.state.value)}
-                  </InputAdornment>
-                ),
-              }}
-            />
-            {this.state.pickerIsOpen ? (
-              <ClickAwayListener onClickAway={this.closePicker}>
-                <div style={style.pickerContainer}>
-                  <SketchPicker
-                    color={this.state.value}
-                    onChangeComplete={this.handleChange}
-                    onSwatchHover={this.handleChange}
-                    disableAlpha={true}
-                  />
-                </div>
-              </ClickAwayListener>
-            ) : null}
-            <FormHelperText>{this.props.settings.Description}</FormHelperText>
-          </FormControl>
-        )
-      case 'browse':
-        return (
-          <FormControl>
-            <TextField
-              type="text"
-              name={this.props.settings.Name}
-              id={this.props.settings.Name}
-              label={this.props.settings.DisplayName}
-              disabled={true}
-              value={this.state.value}
-              InputProps={{
-                startAdornment: (
-                  <InputAdornment position="start">
-                    {this.props.renderIcon
-                      ? this.props.renderIcon('lens')
-                      : renderIconDefault('lens', this.state.value)}
-                  </InputAdornment>
-                ),
-              }}
-            />
-          </FormControl>
-        )
-      default:
-        return (
-          <div>
-            <label>{this.props.settings.DisplayName}</label>
-          </div>
-        )
-    }
+
+  switch (props.actionName) {
+    case 'edit':
+    case 'new':
+      return (
+        <FormControl>
+          <TextField
+            label={props.settings.DisplayName}
+            type="text"
+            name={props.settings.Name}
+            id={props.settings.Name}
+            placeholder={props.settings.DisplayName}
+            defaultValue={props.settings.DefaultValue}
+            required={props.settings.Compulsory}
+            disabled={props.settings.ReadOnly}
+            value={value}
+            onClick={() => setIsPickerOpen(true)}
+            helperText={props.settings.Description}
+            InputProps={{
+              startAdornment: (
+                <InputAdornment position="start">
+                  {props.renderIcon ? props.renderIcon('lens') : renderIconDefault('lens', value)}
+                </InputAdornment>
+              ),
+            }}
+          />
+          {isPickerOpen ? (
+            <ClickAwayListener onClickAway={() => setIsPickerOpen(false)}>
+              <div style={style.pickerContainer}>
+                <SketchPicker
+                  color={value}
+                  onChangeComplete={handleChange}
+                  onSwatchHover={handleChange}
+                  disableAlpha={true}
+                />
+              </div>
+            </ClickAwayListener>
+          ) : null}
+        </FormControl>
+      )
+    case 'browse':
+    default:
+      return props.fieldValue ? (
+        <FormControl>
+          <TextField
+            type="text"
+            name={props.settings.Name}
+            id={props.settings.Name}
+            label={props.settings.DisplayName}
+            disabled={true}
+            value={value}
+            InputProps={{
+              startAdornment: (
+                <InputAdornment position="start">
+                  {props.renderIcon ? props.renderIcon('lens') : renderIconDefault('lens', value)}
+                </InputAdornment>
+              ),
+            }}
+          />
+        </FormControl>
+      ) : null
   }
 }
