@@ -1,7 +1,7 @@
 /**
  * @module FieldControls
  */
-import React, { Component } from 'react'
+import React, { useState } from 'react'
 import FormControl from '@material-ui/core/FormControl'
 import FormControlLabel from '@material-ui/core/FormControlLabel'
 import FormGroup from '@material-ui/core/FormGroup'
@@ -10,167 +10,90 @@ import FormLabel from '@material-ui/core/FormLabel'
 import InputLabel from '@material-ui/core/InputLabel'
 import MenuItem from '@material-ui/core/MenuItem'
 import Select from '@material-ui/core/Select'
-import { ChoiceFieldSetting, GenericContent } from '@sensenet/default-content-types'
+import { ChoiceFieldSetting } from '@sensenet/default-content-types'
+import { changeJScriptValue } from '../helpers'
 import { ReactClientFieldSetting } from './ClientFieldSetting'
 
 /**
- * Interface for DropDownList state
- */
-export interface DropDownListState<T extends GenericContent, K extends keyof T> {
-  value: T[K]
-}
-/**
  * Field control that represents a Choice field. Available values will be populated from the FieldSettings.
  */
-export class DropDownList<T extends GenericContent, K extends keyof T> extends Component<
-  ReactClientFieldSetting<ChoiceFieldSetting>,
-  DropDownListState<T, K>
-> {
-  /**
-   * constructor
-   * @param {object} props
-   */
-  constructor(props: DropDownList<T, K>['props']) {
-    super(props)
-    /**
-     * @type {object}
-     */
-    this.state = {
-      value:
-        (this.props.content && this.props.content[this.props.settings.Name]) || this.props.settings.DefaultValue || [],
+export const DropDownList: React.FC<ReactClientFieldSetting<ChoiceFieldSetting>> = props => {
+  const getInitialstate = () => {
+    if (!props.fieldValue) {
+      return props.settings.AllowMultiple ? [''] : ''
     }
+    if (!Array.isArray(props.fieldValue)) {
+      return props.settings.AllowMultiple ? [props.fieldValue] : props.fieldValue
+    }
+    return props.fieldValue
   }
-  /**
-   * sets the selected value in the state
-   */
-  public handleChange = (event: React.ChangeEvent<{ name?: string; value: unknown }>) => {
-    this.setState({ value: event.target.value as any })
-    this.props.fieldOnChange && this.props.fieldOnChange(this.props.settings.Name, event.target.value as any)
-  }
-  /**
-   * returns selected options value
-   */
-  public getSelectedValue() {
-    let selected
-    this.props.settings.Options!.map(option => {
-      if (option.Selected) {
-        selected = option.Value
-      }
-    })
-    return selected
+  const [value, setValue] = useState(getInitialstate)
+
+  const handleChange = (event: React.ChangeEvent<{ name?: string | undefined; value: unknown }>) => {
+    setValue(event.target.value as any)
+    props.fieldOnChange && props.fieldOnChange(props.settings.Name, value)
   }
 
-  /**
-   * returns selected options text by its value
-   * @param {any} value
-   */
-  public getTextByValue(value: any) {
-    if (value) {
-      this.props.settings.Options!.map(option => {
-        if (option.Value === value.toString()) {
-          return option.Text
-        }
-      })
-    } else {
-      this.props.settings.Options!.map(option => {
-        if (option.Selected) {
-          return option.Text
-        }
-      })
-    }
-  }
-
-  /**
-   * render
-   * @return {ReactElement} markup
-   */
-  public render() {
-    switch (this.props.actionName) {
-      case 'edit':
-        return (
-          <FormControl fullWidth={true} required={this.props.settings.Compulsory}>
-            <InputLabel htmlFor={this.props.settings.Name as string}>{this.props.settings.DisplayName}</InputLabel>
-            <Select
-              onChange={this.handleChange}
-              inputProps={
-                {
-                  name: this.props.settings.Name,
-                  id: this.props.settings.Name,
-                } as any
-              }
-              value={this.state.value[0]}
-              multiple={this.props.settings.AllowMultiple}
-              autoWidth={true}
-              fullWidth={true}>
-              {this.props.settings.Options!.map(option => {
-                return (
-                  <MenuItem key={option.Value} value={option.Value}>
-                    {option.Text}
-                  </MenuItem>
-                )
-              })}
-            </Select>
-            <FormHelperText>{this.props.settings.Description}</FormHelperText>
-          </FormControl>
-        )
-      case 'new':
-        return (
-          <FormControl fullWidth={true} required={this.props.settings.Compulsory}>
-            <InputLabel htmlFor={this.props.settings.Name as string}>{this.props.settings.DisplayName}</InputLabel>
-            <Select
-              onChange={this.handleChange}
-              inputProps={
-                {
-                  name: this.props.settings.Name,
-                  id: this.props.settings.Name,
-                } as any
-              }
-              value={this.state.value as any}
-              multiple={this.props.settings.AllowMultiple}
-              autoWidth={true}
-              fullWidth={true}>
-              {this.props.settings.Options!.map(option => {
-                return (
-                  <MenuItem key={option.Value} value={option.Value}>
-                    {option.Text}
-                  </MenuItem>
-                )
-              })}
-            </Select>
-            <FormHelperText>{this.props.settings.Description}</FormHelperText>
-          </FormControl>
-        )
-      case 'browse':
-      default: {
-        const value = this.props.content && this.props.content[this.props.settings.Name]
-        return value ? (
-          <FormControl component={'fieldset' as 'div'}>
-            <FormLabel component={'legend' as 'label'}>{this.props.settings.DisplayName}</FormLabel>
-            <FormGroup>
-              {Array.isArray(value) ? (
-                value.map((val: any, index: number) => (
-                  <FormControl component={'fieldset' as 'div'} key={index}>
-                    <FormControlLabel
-                      style={{ marginLeft: 0 }}
-                      label={this.props.settings.Options!.find(item => item.Value === val)!.Text}
-                      control={<span />}
-                      key={val}
-                    />
-                  </FormControl>
-                ))
-              ) : (
-                <FormControl component={'fieldset' as 'div'}>
+  switch (props.actionName) {
+    case 'edit':
+    case 'new':
+      return (
+        <FormControl fullWidth={true} required={props.settings.Compulsory} disabled={props.settings.ReadOnly}>
+          <InputLabel htmlFor={props.settings.Name}>{props.settings.DisplayName}</InputLabel>
+          <Select
+            onChange={handleChange}
+            inputProps={
+              {
+                name: props.settings.Name,
+                id: props.settings.Name,
+              } as any
+            }
+            value={value}
+            name={props.settings.Name}
+            multiple={props.settings.AllowMultiple}
+            autoWidth={true}
+            defaultValue={changeJScriptValue(props.settings.DefaultValue)}
+            fullWidth={true}>
+            {props.settings.Options!.map(option => {
+              return (
+                <MenuItem key={option.Value} value={option.Value}>
+                  {option.Text}
+                </MenuItem>
+              )
+            })}
+          </Select>
+          <FormHelperText>{props.settings.Description}</FormHelperText>
+        </FormControl>
+      )
+    case 'browse':
+    default: {
+      return props.fieldValue ? (
+        <FormControl component={'fieldset' as 'div'}>
+          <FormLabel component={'legend' as 'label'}>{props.settings.DisplayName}</FormLabel>
+          <FormGroup>
+            {Array.isArray(props.fieldValue) ? (
+              props.fieldValue.map((val: any, index: number) => (
+                <FormControl component={'fieldset' as 'div'} key={index}>
                   <FormControlLabel
                     style={{ marginLeft: 0 }}
-                    label={this.props.settings.Options!.find(item => item.Value === value)!.Text}
+                    label={props.settings.Options!.find(item => item.Value === val)!.Text}
                     control={<span />}
+                    key={val}
                   />
                 </FormControl>
-              )}
-            </FormGroup>
-          </FormControl>
-        ) : null
-      }
+              ))
+            ) : (
+              <FormControl component={'fieldset' as 'div'}>
+                <FormControlLabel
+                  style={{ marginLeft: 0 }}
+                  label={props.settings.Options!.find(item => item.Value === props.fieldValue)!.Text}
+                  control={<span />}
+                />
+              </FormControl>
+            )}
+          </FormGroup>
+        </FormControl>
+      ) : null
     }
   }
 }
