@@ -1,19 +1,10 @@
 /**
  * @module FieldControls
  */
-import { Component } from 'react'
-import { ReactClientFieldSetting } from './ClientFieldSetting'
 
 /**
  * Field control that represents obsolete fieldcontrols.
  */
-
-export class EmptyFieldControl extends Component<ReactClientFieldSetting, {}> {
-  /**
-   * render
-   * @return {ReactElement} markup
-   */
-  public render() {
-    return null
-  }
+export const EmptyFieldControl = () => {
+  return null
 }
