@@ -8,15 +8,9 @@ import ListItemSecondaryAction from '@material-ui/core/ListItemSecondaryAction'
 import ListItemText from '@material-ui/core/ListItemText'
 import InsertDriveFile from '@material-ui/icons/InsertDriveFile'
 import { GenericContent } from '@sensenet/default-content-types'
-import React, { Component } from 'react'
+import React from 'react'
 import { renderIconDefault } from '../icon'
 import { isUser } from '../type-guards'
-
-const styles = {
-  icon: {
-    marginRight: 0,
-  },
-}
 
 interface DefaultItemTemplateProps {
   content: GenericContent
@@ -24,70 +18,59 @@ interface DefaultItemTemplateProps {
   add: () => void
   actionName?: 'new' | 'edit' | 'browse'
   readOnly?: boolean
-  repositoryUrl: string
+  repositoryUrl?: string
   multiple: boolean
   renderIcon?: (name: string) => JSX.Element
 }
 
-export class DefaultItemTemplate extends Component<DefaultItemTemplateProps, {}> {
-  public handlRemoveIconClick = (id: number) => {
-    if (this.props.remove) {
-      this.props.remove(id)
-    }
-  }
-  public handleAddIconClick = () => {
-    this.props.add()
-  }
-  /**
-   * render
-   * @return {ReactElement} markup
-   */
-  public render() {
-    const { content, repositoryUrl } = this.props
-    return (
-      <ListItem key={content.Id} button={false}>
-        {content.Type ? (
-          isUser(content) ? (
-            <ListItemAvatar>
-              {
-                <Avatar
-                  alt={content.FullName}
-                  src={content.Avatar && content.Avatar.Url ? `${repositoryUrl}${content.Avatar.Url}` : ''}
-                />
-              }
-            </ListItemAvatar>
-          ) : (
-            <ListItemIcon style={styles.icon}>
-              <Icon>
-                <InsertDriveFile />
-              </Icon>
-            </ListItemIcon>
-          )
-        ) : null}
-        <ListItemText
-          primary={content.DisplayName}
-          style={content.Id < 0 ? { textAlign: 'right', paddingRight: 16 } : { textAlign: 'left' }}
-        />
-        {this.props.actionName && this.props.actionName !== 'browse' && !this.props.readOnly ? (
-          <ListItemSecondaryAction>
-            {content ? (
-              content.Id > 0 && this.props.multiple ? (
-                <IconButton onClick={() => this.handlRemoveIconClick(content.Id)}>
-                  {this.props.renderIcon ? this.props.renderIcon('remove_circle') : renderIconDefault('remove_circle')}
-                </IconButton>
-              ) : content.Id === -1 ? (
-                <IconButton onClick={() => this.handleAddIconClick()}>
-                  {this.props.renderIcon ? this.props.renderIcon('add_circle') : renderIconDefault('add_circle')}
-                </IconButton>
-              ) : (
-                <IconButton onClick={() => this.handleAddIconClick()}>
-                  {this.props.renderIcon ? this.props.renderIcon('refresh') : renderIconDefault('refresh')}
-                </IconButton>
-              )
-            ) : null}
-          </ListItemSecondaryAction>
-        ) : null}
-      </ListItem>
-    )
-  }
+/**
+ * Represents a default renderer for reference grid row
+ */
+export const DefaultItemTemplate: React.FC<DefaultItemTemplateProps> = props => {
+  const { content, repositoryUrl } = props
+  return (
+    <ListItem key={content.Id} button={false}>
+      {content.Type ? (
+        isUser(content) ? (
+          <ListItemAvatar>
+            {
+              <Avatar
+                alt={content.FullName}
+                src={
+                  content.Avatar && content.Avatar.Url && repositoryUrl ? `${repositoryUrl}${content.Avatar.Url}` : ''
+                }
+              />
+            }
+          </ListItemAvatar>
+        ) : (
+          <ListItemIcon style={{ marginRight: 0 }}>
+            <Icon>
+              <InsertDriveFile />
+            </Icon>
+          </ListItemIcon>
+        )
+      ) : null}
+      <ListItemText
+        primary={content.DisplayName}
+        style={content.Id < 0 ? { textAlign: 'right', paddingRight: 16 } : { textAlign: 'left' }}
+      />
+      {props.actionName && props.actionName !== 'browse' && !props.readOnly ? (
+        <ListItemSecondaryAction>
+          {content ? (
+            content.Id > 0 && props.multiple ? (
+              <IconButton onClick={() => props.remove && props.remove(content.Id)}>
+                {props.renderIcon ? props.renderIcon('remove_circle') : renderIconDefault('remove_circle')}
+              </IconButton>
+            ) : (
+              <IconButton onClick={() => props.add()}>
+                {props.renderIcon
+                  ? props.renderIcon(content.Id === -1 ? 'add_circle' : 'refresh')
+                  : renderIconDefault(content.Id === -1 ? 'add_circle' : 'refresh')}
+              </IconButton>
+            )
+          ) : null}
+        </ListItemSecondaryAction>
+      ) : null}
+    </ListItem>
+  )
 }
