@@ -4,11 +4,18 @@ import ListItemIcon from '@material-ui/core/ListItemIcon'
 import ListItemText from '@material-ui/core/ListItemText'
 import { ODataParams, Repository } from '@sensenet/client-core'
 import { Folder, GenericContent, User } from '@sensenet/default-content-types'
-import { ListPickerComponent } from '@sensenet/pickers-react'
-import React, { Component } from 'react'
+import { useListPicker } from '@sensenet/pickers-react'
+import React, { useEffect, useState } from 'react'
+import List from '@material-ui/core/List'
+import Fade from '@material-ui/core/Fade'
+import CircularProgress from '@material-ui/core/CircularProgress'
 import { renderIconDefault } from '../icon'
 
 const DEFAULT_AVATAR_PATH = '/Root/Sites/Default_Site/demoavatars/Admin.png'
+const styles: { [index: string]: React.CSSProperties } = {
+  uploadContainer: { minHeight: 50, position: 'relative' },
+  loaderContainer: { display: 'flex', alignItems: 'center', justifyContent: 'center' },
+}
 
 interface ReferencePickerProps {
   change?: () => void
@@ -16,88 +23,113 @@ interface ReferencePickerProps {
   repository: Repository
   path: string
   allowedTypes?: string[]
-  selected: any[]
+  selected: GenericContent[]
   renderIcon?: (name: string) => JSX.Element
 }
-interface ReferencePickerState {
-  path: string
+
+const createTypeFilterString = (allowedTypes: string[]) => {
+  let filterString = "(isOf('Folder') and not isOf('SystemFolder'))"
+  allowedTypes.map((typeName: string) => {
+    if (typeName !== 'Folder') {
+      filterString += ` or isOf('${typeName}')`
+    }
+  })
+  return filterString
 }
 
-export class ReferencePicker extends Component<ReferencePickerProps, ReferencePickerState> {
-  public state: ReferencePickerState = {
-    path: this.props.path,
-  }
-
-  public onNavigation = (path: string) => {
-    this.props.change && this.props.change()
-    this.setState({ path })
-  }
-
-  public onSelectionChanged = (content: GenericContent) => {
-    if (this.props.allowedTypes && this.props.allowedTypes.indexOf(content.Type) > -1) {
-      this.props.select(content)
-    }
-  }
-
-  public createTypeFilterString = (allowedTypes: string[]) => {
-    let filterString = "(isOf('Folder') and not isOf('SystemFolder'))"
-    allowedTypes.map((typeName: string) => {
-      if (typeName !== 'Folder') {
-        filterString += ` or isOf('${typeName}')`
-      }
-    })
-    return filterString
-  }
-
-  private pickerItemOptions: ODataParams<Folder> = {
+/**
+ * Represents a reference picker component
+ */
+export const ReferencePicker: React.FC<ReferencePickerProps> = props => {
+  const pickerItemOptions: ODataParams<Folder> = {
     select: ['DisplayName', 'Path', 'Id', 'Children/IsFolder', 'IsFolder', 'Avatar', 'Icon'] as any,
     expand: ['Children'] as any,
-    filter: this.props.allowedTypes
-      ? this.createTypeFilterString(this.props.allowedTypes)
+    filter: props.allowedTypes
+      ? createTypeFilterString(props.allowedTypes)
       : "(isOf('Folder') and not isOf('SystemFolder'))",
     metadata: 'no',
     orderby: 'DisplayName',
   }
 
-  public iconName = (isFolder?: boolean) => {
+  const { items, setSelectedItem, navigateTo, isLoading, error } = useListPicker<GenericContent>({
+    repository: props.repository,
+    currentPath: props.path,
+    itemsODataOptions: pickerItemOptions,
+  })
+  const [showLoading, setShowLoading] = useState(false)
+
+  const onClickHandler = (_e: React.MouseEvent, node: GenericContent) => {
+    setSelectedItem(node)
+    if (props.allowedTypes && props.allowedTypes.indexOf(node.Type) > -1) {
+      props.select(node)
+    }
+  }
+
+  // Wait to show spinner to prevent content jumping
+  useEffect(() => {
+    const timer = window.setTimeout(() => {
+      setShowLoading(isLoading)
+    }, 800)
+    return () => {
+      window.clearTimeout(timer)
+    }
+  }, [isLoading])
+
+  const iconName = (isFolder?: boolean) => {
     if (isFolder == null) {
       return 'arrow_upward'
     }
     return isFolder ? 'folder' : 'insert_drive_file'
   }
 
-  public renderItem = (node: GenericContent) => (
-    <ListItem button={true} selected={this.props.selected.findIndex(content => node.Id === content.Id) > -1}>
-      <ListItemIcon style={{ margin: 0 }}>
-        {node.Type === 'User' ? (
-          <Avatar
-            alt={node.DisplayName}
-            src={
-              (node as User).Avatar
-                ? `${this.props.repository.configuration.repositoryUrl}${(node as User).Avatar!.Url}`
-                : DEFAULT_AVATAR_PATH
-            }
-          />
-        ) : this.props.renderIcon ? (
-          this.props.renderIcon(this.iconName(node.IsFolder))
-        ) : (
-          renderIconDefault(this.iconName(node.IsFolder))
-        )}
-      </ListItemIcon>
-      <ListItemText primary={node.DisplayName} />
-    </ListItem>
-  )
-
-  public render() {
+  if (showLoading) {
     return (
-      <ListPickerComponent
-        onSelectionChanged={this.onSelectionChanged}
-        onNavigation={this.onNavigation}
-        repository={this.props.repository}
-        currentPath={this.state.path}
-        itemsODataOptions={this.pickerItemOptions}
-        renderItem={this.renderItem}
-      />
+      <div style={styles.loaderContainer}>
+        <Fade
+          in={showLoading}
+          style={{
+            transitionDelay: showLoading ? '800ms' : '0ms',
+          }}
+          unmountOnExit={true}>
+          <CircularProgress />
+        </Fade>
+      </div>
     )
   }
+
+  if (error) {
+    return <p>{error.message}</p>
+  }
+
+  return (
+    <List>
+      {items &&
+        items.map(node => (
+          <ListItem
+            key={node.Id}
+            button={true}
+            onClick={e => onClickHandler(e, node)}
+            onDoubleClick={() => navigateTo(node)}
+            selected={props.selected.some(c => c.Id === node.Id)}>
+            <ListItemIcon style={{ margin: 0 }}>
+              {node.Type === 'User' ? (
+                <Avatar
+                  alt={node.DisplayName}
+                  src={
+                    (node as User).Avatar
+                      ? `${props.repository.configuration.repositoryUrl}${(node as User).Avatar!.Url}`
+                      : DEFAULT_AVATAR_PATH
+                  }
+                />
+              ) : props.renderIcon ? (
+                props.renderIcon(iconName(node.IsFolder))
+              ) : (
+                renderIconDefault(iconName(node.IsFolder))
+              )}
+            </ListItemIcon>
+            <ListItemText primary={node.DisplayName} />
+          </ListItem>
+        ))}
+    </List>
+  )
 }
