@@ -1,70 +1,58 @@
 /**
  * @module FieldControls
  */
-import React, { Component } from 'react'
+import React, { useState } from 'react'
 import TextField from '@material-ui/core/TextField'
 import Typography from '@material-ui/core/Typography'
 import { LongTextFieldSetting } from '@sensenet/default-content-types'
+import { changeJScriptValue } from '../helpers'
 import { ReactClientFieldSetting } from './ClientFieldSetting'
-
-/**
- * Interface for Textarea state
- */
-export interface TextareaState {
-  value: string
-}
 
 /**
  * Field control that represents a LongText field. Available values will be populated from the FieldSettings.
  */
-export class Textarea extends Component<ReactClientFieldSetting<LongTextFieldSetting>, TextareaState> {
-  state = {
-    value:
-      (this.props.content && this.props.content[this.props.settings.Name].replace(/<[^>]*>/g, '')) ||
-      this.props.settings.DefaultValue ||
-      '',
-  }
-  /**
-   * handle change event on an input
-   * @param {SytheticEvent} event
-   */
-  public handleChange = (event: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement | HTMLSelectElement>) => {
-    this.setState({ value: event.target.value })
-    this.props.fieldOnChange && this.props.fieldOnChange(this.props.settings.Name, event.target.value)
+export const Textarea: React.FC<ReactClientFieldSetting<LongTextFieldSetting>> = props => {
+  const initialState =
+    (props.fieldValue && props.fieldValue.replace(/<[^>]*>/g, '')) ||
+    changeJScriptValue(props.settings.DefaultValue) ||
+    ''
+  const [value, setValue] = useState(initialState)
+
+  const handleChange = (event: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement | HTMLSelectElement>) => {
+    setValue(event.target.value)
+    props.fieldOnChange && props.fieldOnChange(props.settings.Name, event.target.value)
   }
 
-  public render() {
-    switch (this.props.actionName) {
-      case 'edit':
-      case 'new':
-        return (
-          <TextField
-            onChange={this.handleChange}
-            name={this.props.settings.Name}
-            id={this.props.settings.Name}
-            label={this.props.settings.DisplayName}
-            placeholder={this.props.settings.DisplayName}
-            value={this.state.value}
-            required={this.props.settings.Compulsory}
-            disabled={this.props.settings.ReadOnly}
-            multiline={true}
-            fullWidth={true}
-          />
-        )
-      case 'browse':
-      default:
-        return this.props.content &&
-          this.props.content[this.props.settings.Name] &&
-          this.props.content[this.props.settings.Name].length > 0 ? (
-          <div>
-            <Typography variant="caption" gutterBottom={true}>
-              {this.props.settings.DisplayName}
-            </Typography>
-            <Typography variant="body1" gutterBottom={true}>
-              {this.props.content[this.props.settings.Name]}
-            </Typography>
-          </div>
-        ) : null
-    }
+  switch (props.actionName) {
+    case 'edit':
+    case 'new':
+      return (
+        <TextField
+          onChange={handleChange}
+          name={props.settings.Name}
+          id={props.settings.Name}
+          label={props.settings.DisplayName}
+          defaultValue={changeJScriptValue(props.settings.DefaultValue)}
+          placeholder={props.settings.DisplayName}
+          value={value}
+          required={props.settings.Compulsory}
+          disabled={props.settings.ReadOnly}
+          multiline={true}
+          fullWidth={true}
+          helperText={props.settings.Description}
+        />
+      )
+    case 'browse':
+    default:
+      return props.fieldValue ? (
+        <div>
+          <Typography variant="caption" gutterBottom={true}>
+            {props.settings.DisplayName}
+          </Typography>
+          <Typography variant="body1" gutterBottom={true}>
+            {props.fieldValue}
+          </Typography>
+        </div>
+      ) : null
   }
 }
