@@ -11,7 +11,7 @@ export * from './TimePicker'
 export * from './DropDownList'
 export * from './RadioButtonGroup'
 export * from './ShortText'
-export * from './TagsInput/TagsInput'
+export * from './TagsInput'
 export * from './Textarea'
 export * from './RichTextEditor'
 export * from './Name'
