@@ -3,12 +3,6 @@
  * @description General classes, modules and methods for sn-controls-react package
  */
 
-import { ReactClientFieldConfig } from './ReactClientFieldConfig'
-import { reactControlMapper } from './ReactControlMapper'
-
-export { ReactClientFieldConfig, reactControlMapper }
-
+export * from './ReactControlMapper'
 export * from './fieldcontrols'
-export * from './viewcontrols/EditView'
-export * from './viewcontrols/NewView'
-export { BrowseView } from './viewcontrols/BrowseView'
+export * from './viewcontrols'
