@@ -20,7 +20,7 @@ export interface BrowseViewProps {
 /**
  * View Control for browsing a Content, works with a single Content and based on the ReactControlMapper
  */
-export function BrowseView(props: BrowseViewProps) {
+export const BrowseView: React.FC<BrowseViewProps> = props => {
   const controlMapper = reactControlMapper(props.repository)
   const schema = controlMapper.getFullSchemaForContentType(props.content.Type, 'browse')
 
@@ -39,6 +39,7 @@ export function BrowseView(props: BrowseViewProps) {
                   actionName: 'browse',
                   settings: field.fieldSettings,
                   content: props.content,
+                  fieldValue: (props.content as any)[field.fieldSettings.Name],
                   renderIcon: props.renderIcon,
                   repository: props.repository,
                 },
