@@ -25,7 +25,8 @@ export interface ReferenceFieldProps<T> {
   fieldSetting: ReferenceFieldSetting
   defaultValueIdOrPath?: string | number
   fetchItems: (fetchQuery: Query<T>) => Promise<T[]>
-  onQueryChange: (key: string, query: Query<T>) => void
+  onQueryChange?: (key: string, query: Query<T>) => void
+  onChange?: (item: T) => void
   renderSuggestion?: RenderSuggestion<T>
   autoSuggestProps?: Partial<AutosuggestProps<T>>
 }
@@ -99,10 +100,11 @@ export class ReferenceField<T extends GenericContent = GenericContent> extends R
           inputValue: item.DisplayName || item.Name,
           selected: item,
         })
-        this.props.onQueryChange(
-          this.props.fieldKey || this.props.fieldName.toString(),
-          new Query(q => q.equals(this.props.fieldName, item.Id)),
-        )
+        this.props.onQueryChange &&
+          this.props.onQueryChange(
+            this.props.fieldKey || this.props.fieldName.toString(),
+            new Query(q => q.equals(this.props.fieldName, item.Id)),
+          )
       }
     }
   }
@@ -144,14 +146,16 @@ export class ReferenceField<T extends GenericContent = GenericContent> extends R
   }
 
   private handleSelect(item: T) {
+    this.props.onChange && this.props.onChange(item)
     this.setState({
       inputValue: item.DisplayName || item.Name,
       selected: item,
     })
-    this.props.onQueryChange(
-      this.props.fieldKey || this.props.fieldName.toString(),
-      new Query(q => q.equals(this.props.fieldName, item.Id)),
-    )
+    this.props.onQueryChange &&
+      this.props.onQueryChange(
+        this.props.fieldKey || this.props.fieldName.toString(),
+        new Query(q => q.equals(this.props.fieldName, item.Id)),
+      )
   }
 
   /**
